Made process identifiers generic.
Previously, we had "type proc = int". Now, there is a separate
module for process identifiers, and all operations on process
identifiers now go via that module. The module still defines
process identifiers as integers, but the plan is for there soon
to be several implementations of that module, so that scoped
models can define a process identifier as a tuple.