Location kinds now work in CPP11. One can refer to the sets "atomicloc" "nonatomicloc" and "mutexloc" of events that access atomic/non-atomic/mutex locations, respectively.
I'm not overjoyed with my solution to this. The location-kind map
now infiltrates rather a lot of places, for instance, every
program parser (PPCParser, ARMParser, etc.) must now return a
location-kind map (which is, by default, the empty map). It would be
nicer if these parsers didn't have to know anything about
location-kind maps. I'll keep pondering about whether there's a
better way to do this, but it's tricky because it's quite hard to
modify code in the lib directory without breaking the other tools
in the diy suite.